#459 Moves objectToKeyValues to utils and adds unit test
